@@ -120,111 +124,120 @@
 		"ng-dist",
 		"node_modules"
 	],
-	"devDependencies": {
-		"@angular-devkit/build-angular": "^0.802.2",
-		"@angular/cli": "^8.2.2",
-		"@angular/compiler": "^8.2.2",
-		"@angular/compiler-cli": "^8.2.2",
-		"@angular/language-service": "^8.2.2",
-		"@types/echarts": "^4.1.10",
-		"@types/jasmine": "^3.4.0",
-		"@types/jasminewd2": "~2.0.6",
-		"@types/node": "^12.7.2",
-		"@webcomponents/custom-elements": "^1.2.4",
-		"ajv": "^6.10.2",
-		"angular2-text-mask": "^9.0.0",
-		"angular2-toaster": "^8.0.0",
-		"bootstrap": "^4.3.1",
-		"codelyzer": "~5.1.0",
-		"core-js": "^2.6.9",
-		"echarts": "^4.2.1",
-		"electron": "^6.0.2",
-		"electron-builder": "^21.2.0",
-		"electron-reload": "^1.5.0",
-		"fuse.js": "^3.4.5",
-		"hammerjs": "^2.0.8",
-		"jquery": "^3.4.0",
-		"lodash-es": "^4.17.15",
-		"lodash.get": "^4.4.2",
-		"lodash.set": "^4.3.2",
-		"moment": "^2.24.0",
-		"ng-lottie": "^0.3.2",
-		"ngx-echarts": "^4.2.1",
-		"ngx-electron": "^2.1.1",
-		"ngx-json-viewer": "^2.4.0",
-		"ngx-markdown": "^8.1.0",
-		"ngx-order-pipe": "^2.0.3",
-		"ngx-pagination": "^4.1.0",
-		"npm-run-all": "^4.1.5",
-		"popper.js": "^1.15.0",
-		"primeng": "^8.0.1",
-		"protractor": "^5.4.2",
-		"rxjs": "^6.5.2",
-		"rxjs-compat": "^6.5.2",
-		"text-encoding-shim": "latest",
-		"text-mask-addons": "^3.8.0",
-		"ts-node": "^8.3.0",
-		"tslint": "^5.18.0",
-		"typescript": "3.4.5",
-		"wait-on": "^3.3.0",
-		"webpack": "^4.39.2",
-		"webpack-dev-server": "^3.8.0",
-		"zone.js": "^0.9.1"
-	},
 	"dependencies": {
-		"@angular/animations": "^8.2.2",
-		"@angular/cdk": "^8.1.3",
-		"@angular/common": "^8.2.2",
-		"@angular/core": "^8.2.2",
-		"@angular/flex-layout": "^8.0.0-beta.26",
-		"@angular/forms": "^8.2.2",
-		"@angular/http": "^8.0.0-beta.10",
-		"@angular/material": "^8.1.3",
-		"@angular/material-moment-adapter": "^8.1.3",
-		"@angular/platform-browser": "^8.2.2",
-		"@angular/platform-browser-dynamic": "^8.2.2",
-		"@angular/router": "^8.2.2",
-		"@clr/angular": "^2.1.1",
-		"@clr/icons": "^2.1.1",
-		"@clr/ui": "^2.1.1",
-		"@fortawesome/angular-fontawesome": "^0.5.0",
-		"@fortawesome/fontawesome-svg-core": "^1.2.21",
-		"@fortawesome/free-brands-svg-icons": "^5.9.0",
-		"@fortawesome/free-regular-svg-icons": "^5.9.0",
-		"@fortawesome/free-solid-svg-icons": "^5.9.0",
-		"@fortawesome/pro-light-svg-icons": "^5.9.0",
-		"@fortawesome/pro-regular-svg-icons": "^5.9.0",
-		"@fortawesome/pro-solid-svg-icons": "^5.9.0",
-		"@ledgerhq/hw-transport": "latest",
-		"@types/crypto-js": "^3.1.43",
-		"angular-schema-form": "^0.8.13",
-		"angular7-json-schema-form": "^1.0.4",
+		"@ajsf/core": "0.1.3",
+		"@ajsf/material": "0.1.3",
+		"@angular/animations": "9.0.5",
+		"@angular/cdk": "9.1.1",
+		"@angular/common": "9.0.5",
+		"@angular/core": "9.0.5",
+		"@angular/flex-layout": "9.0.0-beta.29",
+		"@angular/forms": "9.0.5",
+		"@angular/material": "9.1.1",
+		"@angular/material-moment-adapter": "9.1.1",
+		"@angular/platform-browser": "9.0.5",
+		"@angular/platform-browser-dynamic": "9.0.5",
+		"@angular/router": "9.0.5",
+		"@clr/angular": "3.0.0",
+		"@clr/core": "3.0.0",
+		"@clr/icons": "3.0.0",
+		"@clr/ui": "3.0.0",
+		"@fortawesome/angular-fontawesome": "^0.6.0",
+		"@fortawesome/fontawesome-svg-core": "^1.2.27",
+		"@fortawesome/free-brands-svg-icons": "^5.12.1",
+		"@fortawesome/free-regular-svg-icons": "^5.12.1",
+		"@fortawesome/free-solid-svg-icons": "^5.12.1",
+		"@fortawesome/pro-duotone-svg-icons": "^5.12.1",
+		"@fortawesome/pro-light-svg-icons": "^5.12.1",
+		"@fortawesome/pro-regular-svg-icons": "^5.12.1",
+		"@fortawesome/pro-solid-svg-icons": "^5.12.1",
+		"@ledgerhq/hw-transport": "^5.9.0",
+		"@ledgerhq/hw-transport-node-hid": "^5.10.0",
+		"ajv": "^6.12.0",
+		"angular2-text-mask": "9.0.0",
+		"angular2-toaster": "8.0.0",
 		"animate.css": "^3.7.2",
 		"asn1-ber": "^1.0.9",
 		"auto-launch": "^5.0.5",
+		"bcrypt": "^4.0.1",
 		"bip32-path": "^0.4.2",
 		"body-parser": "^1.19.0",
-		"chart.js": "^2.8.0",
+		"bootstrap": "^4.4.1",
+		"chart.js": "^2.9.3",
 		"cors": "^2.8.5",
-		"crypto-js": "^3.1.9-1",
-		"electron-context-menu": "^0.14.0",
-		"eosjs": "^20.0.0",
+		"crypto-js": "^4.0.0",
+		"echarts": "^4.6.0",
+		"electron-context-menu": "^0.16.0",
+		"eosjs": "^21.0.1-rc2",
 		"express": "^4.17.1",
 		"fcbuffer": "^2.2.2",
+		"fuse.js": "^3.4.6",
+		"hammerjs": "^2.0.8",
 		"isomorphic-fetch": "^2.2.1",
-		"keytar": "^4.13.0",
-		"loaderjs": "^1.0.0",
+		"jquery": "^3.4.1",
+		"keytar": "^5.4.0",
+		"lodash-es": "^4.17.15",
+		"lodash.filter": "4.6.0",
+		"lodash.get": "^4.4.2",
+		"lodash.map": "4.6.0",
+		"lodash.set": "4.3.2",
+		"lottie-web": "^5.6.6",
 		"menu": "^0.2.5",
+		"moment": "^2.24.0",
+		"ngx-echarts": "^4.2.2",
+		"ngx-electron": "^2.2.0",
+		"ngx-json-viewer": "^2.4.0",
+		"ngx-lottie": "^5.3.2",
+		"ngx-markdown": "^9.0.0",
+		"ngx-order-pipe": "^2.0.4",
+		"ngx-pagination": "^5.0.0",
 		"node-fetch": "^2.6.0",
-		"node-pre-gyp": "^0.13.0",
+		"node-pre-gyp": "^0.14.0",
+		"node-sass": "^4.13.1",
 		"node-schedule": "^1.3.2",
-		"open": "^6.4.0",
-		"portfinder": "^1.0.22",
+		"open": "^7.0.2",
+		"popper.js": "^1.16.1",
+		"portfinder": "^1.0.25",
 		"primeicons": "^2.0.0",
-		"quill": "2.0.0-dev.3",
-		"socket.io": "^2.2.0",
-		"socket.io-client": "^2.2.0",
-		"tar": "^4.4.10",
-		"tslib": "^1.10.0"
+		"primeng": "^9.0.0",
+		"quill": "^2.0.0-dev.3",
+		"socket.io": "^2.3.0",
+		"socket.io-client": "^2.3.0",
+		"sudo-prompt": "^9.1.1",
+		"tar": "^6.0.1",
+		"text-mask-addons": "3.8.0",
+		"tslib": "^1.11.1",
+		"@webcomponents/webcomponentsjs": "^2.0.0"
+	},
+	"devDependencies": {
+		"@angular-devkit/build-angular": "0.900.5",
+		"@angular/cli": "9.0.5",
+		"@angular/compiler": "9.0.5",
+		"@angular/compiler-cli": "9.0.5",
+		"@angular/language-service": "9.0.5",
+		"@electron/typescript-definitions": "8.6.6",
+		"@types/crypto-js": "3.1.43",
+		"@types/echarts": "4.4.3",
+		"@types/express": "4.17.3",
+		"@types/node": "13.7.7",
+		"@webcomponents/custom-elements": "^1.4.0",
+		"codelyzer": "^5.1.2",
+		"core-js": "^3.6.4",
+		"electron": "^8.1.0",
+		"electron-builder": "^22.4.0",
+		"electron-builder-notarize": "^1.1.2",
+		"electron-reload": "^1.5.0",
+		"eslint": "^6.8.0",
+		"npm-run-all": "^4.1.5",
+		"protractor": "^5.4.3",
+		"rxjs": "^6.5.4",
+		"rxjs-compat": "^6.5.4",
+		"text-encoding-shim": "^1.0.5",
+		"ts-node": "^8.6.2",
+		"typescript": "3.7.5",
+		"wait-on": "^4.0.1",
+		"webpack": "^4.42.0",
+		"webpack-dev-server": "^3.10.3",
+		"zone.js": "^0.10.2"
 	}
 }
