@@ -1,314 +1,366 @@
-import {Injectable} from '@angular/core';
+import {EventEmitter, Injectable} from '@angular/core';
 import {AccountsService} from './accounts.service';
-import {EOSJSService} from './eosjs.service';
 import {Router} from '@angular/router';
 
-import * as Eos from '../../assets/eos.js';
 import {BehaviorSubject} from 'rxjs';
-import {CryptoService} from './crypto.service';
+import {CryptoService} from './crypto/crypto.service';
 import {VotingService} from './voting.service';
-import {Eosjs2Service} from './eosjs2.service';
-
+import {Eosjs2Service} from './eosio/eosjs2.service';
+import {HttpClient} from '@angular/common/http';
+import {localConfig} from '../../config';
+import {environment} from '../../environments/environment';
 
 export interface Endpoint {
-	url: string;
-	owner: string;
-	latency: number;
-	filters: any[];
-	chain: string;
+    url: string;
+    owner: string;
+    latency: number;
+    filters: any[];
+    chain: string;
 }
 
-@Injectable ( {
-	providedIn: 'root'
-} )
+const defaultCompilerIds = {
+    DEFAULT: 'aca376f206b8fc25a6ed44dbdc66547c36c6c33e3a119ffbeaef943642f0e906',
+    LIBERLAND: 'cc7d69ef6216ba33be85e9b256fbfbad4e103c14e0f115b281b2f954838c463a'
+};
+
+@Injectable({
+    providedIn: 'root',
+})
 export class NetworkService {
 
-	publicEndpoints: Endpoint[];
-	eos: any;
+    eos: any;
+    public mainnetId: string;
+    public: string;
+    validEndpoints: Endpoint[];
+    status: string;
+    connectionTimeout: any;
+    selectedEndpoint = new BehaviorSubject<Endpoint>(null);
+    networkingReady = new BehaviorSubject<boolean>(false);
 
-	public mainnetId: string;
+    connected = false;
+    lastEndpoint = '';
+    autoMode = false;
 
-	public: string;
-	genesistx = 'ad77575a8b4f52e477682e712b1cbd884299468db6a94d909f90c6961cea9b02';
-	voteref = 'b23f537e8ab29fbcec8b533081ef7e12b146899ca42a3fc9eb608258df9983d9';
-	accountez = 'EOS7WdCcva3WtsJRckJWodnHLof5B7qwAyfJSaMZmfn7Dgn6TQDBu';
-	txrefBlock = 191;
-	voterefBlock = 572278;
-	baseConfig = {
-		httpEndpoint: '' ,
-		expireInSeconds: 60 ,
-		broadcast: true ,
-		debug: false ,
-		sign: true ,
-		chainId: ''
-	};
-	validEndpoints: Endpoint[];
-	status: string;
-	connectionTimeout: any;
-	selectedEndpoint = new BehaviorSubject<Endpoint> ( null );
-	networkingReady = new BehaviorSubject<boolean> ( false );
+    public activeChain = null;
+    defaultChains: any[];
+    selectGroup: any[];
+    isStarting = false;
 
-	connected = false;
-	lastEndpoint = '';
-	autoMode = false;
+    customChainModal = false;
+    events: EventEmitter<any>;
 
-	public activeChain = null;
-	defaultChains: any[];
-	selectGroup: any[];
+    static groupBy(list, keyGetter) {
+        const map = new Map();
+        list.forEach((item) => {
+            const key = keyGetter(item);
+            const collection = map.get(key);
+            if (!collection) {
+                map.set(key, [item]);
+            } else {
+                collection.push(item);
+            }
+        });
+        return map;
+    }
 
-	static groupBy(list , keyGetter) {
-		const map = new Map ();
-		list.forEach ( (item) => {
-			const key = keyGetter ( item );
-			const collection = map.get ( key );
-			if (!collection) {
-				map.set ( key , [ item ] );
-			} else {
-				collection.push ( item );
-			}
-		} );
-		return map;
-	}
+    constructor(
+        private http: HttpClient,
+        private eosjs: Eosjs2Service,
+        private router: Router,
+        private aService: AccountsService,
+        private voting: VotingService,
+        private crypto: CryptoService,
+    ) {
+        this.initChainsConfig();
+        this.validEndpoints = [];
+        this.status = '';
+        this.connectionTimeout = null;
+        this.events = new EventEmitter(true);
+    }
 
-	constructor(
-		private eosjs: EOSJSService ,
-		private eosjs2: Eosjs2Service ,
-		private router: Router ,
-		private aService: AccountsService ,
-		private voting: VotingService ,
-		private crypto: CryptoService ,
-		// private ledger: LedgerHWService
-	) {
-		const configSimpleos = JSON.parse ( localStorage.getItem ( 'configSimpleos' ) );
-		this.defaultChains = configSimpleos[ 'config' ][ 'chains' ];
-		const groupChain = NetworkService.groupBy ( this.defaultChains , chain => chain.network );
-		const mainnet = groupChain.get ( 'MAINNET' );
-		const testenet = groupChain.get ( 'TESTNET' );
-		this.selectGroup = [ {'name': 'MAINNETS' , 'chains': mainnet} , {'name': 'TESTNETS' , 'chains': testenet} ];
+    openCustomChainModal() {
+        console.log('open custom chain modal');
+        this.customChainModal = true;
+    }
 
-		this.activeChain = this.aService.activeChain;
-		this.validEndpoints = [];
-		this.status = '';
-		this.connectionTimeout = null;
-	}
+    createGroups() {
+        const groupChain = NetworkService.groupBy(this.defaultChains, chain => chain.network);
+        this.selectGroup = [];
+        const mainnetList = groupChain.get('MAINNET');
+        if (mainnetList) {
+            this.selectGroup.push({
+                'name': 'MAINNETS',
+                'chains': mainnetList,
+            });
+        }
+        const testnetList = groupChain.get('TESTNET');
+        if (testnetList) {
+            this.selectGroup.push({
+                'name': 'TESTNETS',
+                'chains': testnetList,
+            });
+        }
+    }
 
-	connect(automatic: boolean) {
-		// console.log('analyzing endpoints...');
-		this.autoMode = automatic;
-		this.status = '';
-		this.mainnetId = '';
-		this.aService.activeChain = this.activeChain;
-		this.mainnetId = this.activeChain[ 'id' ];
-		this.networkingReady.next ( false );
-		const pQueue = [];
-		this.connected = false;
-		this.activeChain[ 'endpoints' ].forEach ( (apiNode) => {
-			pQueue.push ( this.apiCheck ( apiNode ) );
-		} );
-		Promise.all ( pQueue ).then ( () => {
-			this.extractValidNode ();
-		} );
-		this.startTimeout ();
-	}
+    initChainsConfig() {
 
-	changeChain(chainId) {
-		this.activeChain = this.defaultChains.find ( (chain) => chain.id === chainId );
-		if (this.activeChain) {
-			this.aService.activeChain = this.activeChain;
-			this.aService.accounts = [];
-			this.voting.clearMap ();
-			this.voting.clearLists ();
-			this.voting.initList = false;
-			this.aService.lastAccount = null;
-			localStorage.setItem ( 'simplEOS.activeChainID' , this.activeChain.id );
-			this.connect ( false );
-			console.log ( 'Network switched to: ' + this.activeChain[ 'name' ] );
-		}
-	}
+        let configSimpleos = JSON.parse(localStorage.getItem('configSimpleos'));
+        if (!configSimpleos) {
+            console.log('failed to load updated config');
+            configSimpleos = {
+                config: localConfig
+            };
+        }
 
-	startTimeout() {
-		if (!this.connectionTimeout) {
-			this.connectionTimeout = setTimeout ( () => {
-				console.log ( 'Timeout!' );
-				if (!this.networkingReady.getValue ()) {
-					this.status = 'timeout';
-					clearTimeout ( this.connectionTimeout );
-					this.networkingReady.next ( false );
-					this.connectionTimeout = null;
-				}
-			} , 10000 );
-		}
-	}
+        if (environment.COMPILERVERSION === 'DEFAULT') {
+            this.defaultChains = configSimpleos.config.chains;
+        } else {
+            this.defaultChains = configSimpleos.config.chains.filter(chain => {
+                return chain.name.startsWith(environment.COMPILERVERSION);
+            });
+        }
 
-	extractValidNode() {
-		this.validEndpoints = [];
-		this.activeChain.endpoints.forEach ( (apiNode) => {
-			if (apiNode.latency > 0 && apiNode.latency < 1200) {
-				this.validEndpoints.push ( apiNode );
-			}
-		} );
-		this.selectEndpoint ();
-	}
+        // load custom chains
+        try {
+            const customChains = JSON.parse(localStorage.getItem('custom_chains'));
+            if (customChains) {
+                for (const chain of customChains) {
+                    this.defaultChains.push(chain);
+                    console.log(`Added ${chain.name} with id ${chain.id}`);
+                }
+            }
+        } catch (e) {
+            console.log(e);
+        }
 
-	selectEndpoint() {
-		let latency = 2000;
-		if (this.connected === false) {
-			this.validEndpoints.forEach ( (node) => {
-				if (node.latency < latency && node.latency > 1) {
-					latency = node.latency;
-					this.selectedEndpoint.next ( node );
-				}
-			} );
-			if (this.selectedEndpoint.getValue () === null) {
-				this.networkingReady.next ( false );
-			} else {
-				console.log ( 'Best Server Selected!' , this.selectedEndpoint.getValue ().url );
-				this.startup ( null );
-			}
-		}
-	}
+        this.createGroups();
 
-	selectedEP() {
-		return this.eosjs.baseConfig.httpEndpoint;
-	}
+        const savedChainId = localStorage.getItem('simplEOS.activeChainID');
+        const EOS_MAINNET_ID = 'aca376f206b8fc25a6ed44dbdc66547c36c6c33e3a119ffbeaef943642f0e906';
 
-	filterCheck(server: Endpoint) {
-		console.log ( 'Starting filter check for ' + server.url );
-		const config = this.baseConfig;
-		config.httpEndpoint = server.url;
-		config.chainId = this.mainnetId;
-		const eosCK = Eos ( config );
-		const pq = [];
-		const getkeyAcc = eosCK[ 'getKeyAccounts' ] ( this.accountez ).then ( info => {
+        if (savedChainId) {
+            this.activeChain = this.defaultChains.find((chain) => chain.id === savedChainId);
+            if (!this.activeChain) {
+                console.log('Saved chain not found!');
+                if (environment.COMPILERVERSION === 'DEFAULT') {
+                    this.activeChain = this.defaultChains.find((chain) => chain.id === EOS_MAINNET_ID);
+                    localStorage.setItem('simplEOS.activeChainID', EOS_MAINNET_ID);
+                } else {
+                    this.activeChain = this.defaultChains[0];
+                    localStorage.setItem('simplEOS.activeChainID', this.activeChain.id);
+                }
+            }
+        } else {
+            this.activeChain = this.defaultChains.find((chain) => chain.id === EOS_MAINNET_ID);
+            localStorage.setItem('simplEOS.activeChainID', EOS_MAINNET_ID);
+        }
+        this.aService.activeChain = this.activeChain;
+        this.aService.defaultChains = this.defaultChains;
+        this.aService.init();
+    }
 
-			if (info.length > 0 || info[ 'account_names' ].length > 0) {
-				this.publicEndpoints.find ( ep => ep.url === server.url ).filters.push ( {eosio: 'history'} );
-				return true;
-			} else {
-				console.log ( 'eosio:history filter is disabled on ' + server.url );
-			}
-		} ).catch ( err => {
-			console.log ( err );
-			return false;
-		} );
-		pq.push ( getkeyAcc );
-		return Promise.all ( pq );
-	}
+    async connect(automatic: boolean) {
+        this.autoMode = automatic;
+        this.status = '';
+        this.mainnetId = '';
+        this.aService.activeChain = this.activeChain;
+        this.mainnetId = this.activeChain['id'];
+        this.networkingReady.next(false);
+        const pQueue = [];
+        this.connected = false;
+        this.activeChain['endpoints'].forEach((apiNode) => {
+            pQueue.push(this.apiCheck(apiNode));
+        });
+        Promise.all(pQueue).then(() => {
+            this.extractValidNode();
+        });
+        this.startTimeout();
+        await new Promise(resolve => {
+            const evSub = this.events.asObservable().subscribe(value => {
+                if (value.event === 'connected' || value.event === 'timeout') {
+                    evSub.unsubscribe();
+                    resolve();
+                }
+            });
+        });
+    }
 
-	apiCheck(server: Endpoint) {
-		// console.log('Starting latency check for ' + server.url);
-		return new Promise ( (resolve) => {
-			const config = this.baseConfig;
-			config.httpEndpoint = server.url;
-			config.chainId = this.mainnetId;
-			const eos = Eos ( config );
-			const refTime = new Date ().getTime ();
-			const tempTimer = setTimeout ( () => {
-				server.latency = -1;
-				resolve ();
-			} , 2000 );
-			try {
-				eos[ 'getInfo' ] ( {} , (err) => {
-					if (err) {
-						server.latency = -1;
-					} else {
-						server.latency = ((new Date ().getTime ()) - refTime);
-						// console.log(server.url, server.latency);
-					}
-					clearTimeout ( tempTimer );
-					if (server.latency > 1 && server.latency < 200) {
-						// force quick connection
-						if (this.connected === false) {
-							this.connected = true;
-							this.callStartupConn ( server );
-						}
-					}
-					resolve ();
-				} );
-			} catch (e) {
-				server.latency = -1;
-				resolve ();
-			}
-		} );
-	}
+    async changeChain(chainId) {
+        this.activeChain = this.defaultChains.find((chain) => chain.id === chainId);
+        if (this.activeChain) {
 
-	callStartupConn(server) {
-		if (this.connected === true) {
-			// console.log('fast api detected, connecting to:', server.url);
-			this.selectedEndpoint.next ( server );
-			this.startup ( null );
-		}
-	}
+            // send to landing page if no data was found
+            const lsKey = "simpleos.accounts." + chainId;
+            const savedAccountData = localStorage.getItem(lsKey);
 
-	startup(url) {
-		let endpoint = url;
-		if (!url) {
-			endpoint = this.selectedEndpoint.getValue ().url;
-			// console.log('switcing to saved endpoint:', endpoint);
-		} else {
-			this.status = '';
-			console.log ( 'startup called - url: ' , url );
-		}
-		this.networkingReady.next ( false );
-		this.eosjs.online.next ( false );
-		this.startTimeout ();
-		// prevent double load after quick connection mode
-		if (endpoint !== this.lastEndpoint || this.autoMode === true) {
-			this.eosjs2.initRPC ( endpoint , this.activeChain.id );
-			this.eosjs.init ( endpoint , this.activeChain.id ).then ( (savedAccounts: any) => {
-				// if (this.ledger.isElectron()) {
-				//   this.aService.checkLedgerAccounts().then(() => {
-				//     this.ledger.initListener();
-				//   });
-				// }
-				this.lastEndpoint = endpoint;
-				this.autoMode = false;
-				this.defaultChains.find ( c => c.id === this.activeChain.id ).lastNode = this.lastEndpoint;
+            if (!savedAccountData) {
+                await this.router.navigateByUrl('/');
+            }
 
-				if (savedAccounts) {
-					if (savedAccounts.length > 0) {
-						// console.log('Locading local accounts');
-						this.aService.loadLocalAccounts ( savedAccounts ).then ( () => {
-							if (this.aService.lastAccount) {
-								this.aService.select ( this.aService.accounts.findIndex ( (a) => {
-									return a.name === this.aService.lastAccount;
-								} ) );
-							} else {
-								this.aService.initFirst ();
-							}
-							this.router[ 'navigate' ] ( [ 'dashboard' , 'wallet' ] );
-							// this.voting.forceReload();
-							this.networkingReady.next ( true );
-						} );
-					} else {
-						this.networkingReady.next ( true );
-						if (this.crypto.locked) {
-							console.log ( 'No saved accounts!' );
-							this.router[ 'navigate' ] ( [ '' ] );
-						} else {
-							console.log ( 'No saved accounts!' );
-							this.router[ 'navigate' ] ( [ 'landing' ] );
-						}
-					}
-				}
+            this.aService.activeChain = this.activeChain;
+            this.aService.accounts = [];
+            this.voting.clearMap();
+            this.voting.clearLists();
+            this.voting.initList = false;
+            this.aService.lastAccount = null;
+            localStorage.setItem('simplEOS.activeChainID', this.activeChain.id);
+            await this.connect(false);
+            console.log('Network switched to: ' + this.activeChain['name']);
+        }
+    }
 
-				if (this.connectionTimeout) {
-					clearTimeout ( this.connectionTimeout );
-					this.networkingReady.next ( true );
-					this.connectionTimeout = null;
-				}
-			} ).catch ( (err) => {
-				console.log ( '>>> EOSJS_ERROR: ' , err );
-				this.networkingReady.next ( false );
-			} );
-		} else {
-			if (this.connectionTimeout) {
-				clearTimeout ( this.connectionTimeout );
-				this.networkingReady.next ( true );
-				this.connectionTimeout = null;
-			}
-			this.networkingReady.next ( true );
-		}
-	}
+    startTimeout() {
+        if (!this.connectionTimeout) {
+            this.connectionTimeout = setTimeout(() => {
+                if (!this.networkingReady.getValue()) {
+                    this.status = 'timeout';
+                    clearTimeout(this.connectionTimeout);
+                    this.networkingReady.next(false);
+                    this.events.emit({event: 'timeout'});
+                    this.connectionTimeout = null;
+                    this.isStarting = false;
+                }
+            }, 5000);
+        }
+    }
+
+    extractValidNode() {
+        this.validEndpoints = [];
+        this.activeChain.endpoints.forEach((apiNode) => {
+            if (apiNode.latency > 0 && apiNode.latency < 1200) {
+                this.validEndpoints.push(apiNode);
+            }
+        });
+        this.selectEndpoint();
+    }
+
+    selectEndpoint() {
+        let latency = 2000;
+        if (this.connected === false) {
+            for (const node of this.validEndpoints) {
+                if (node.latency < latency && node.latency > 1) {
+                    latency = node.latency;
+                    this.selectedEndpoint.next(node);
+                }
+            }
+
+            if (this.selectedEndpoint.getValue()) {
+                console.log('Best Server Selected!', this.selectedEndpoint.getValue().url);
+                this.startup(null).catch(console.log);
+            } else {
+                this.networkingReady.next(false);
+            }
+        }
+    }
+
+    selectedEP() {
+        return this.eosjs.baseConfig.httpEndpoint;
+    }
+
+    apiCheck(server: Endpoint) {
+        return new Promise((resolve) => {
+            const refTime = new Date().getTime();
+            const tempTimer = setTimeout(() => {
+                server.latency = -1;
+                resolve();
+            }, 2000);
+            this.http.get(`${server.url}/v1/chain/get_info`).toPromise().then((data: any) => {
+                if (data['chain_id'] === this.activeChain.id) {
+                    server.latency = ((new Date().getTime()) - refTime);
+                    clearTimeout(tempTimer);
+                    if (server.latency > 1 && server.latency < 200) {
+                        if (this.connected === false) {
+                            this.connected = true;
+                            this.callStartupConn(server);
+                        }
+                    }
+                    resolve();
+                } else {
+                    console.log(`API ${server.url} is serving a different chain id!
+                        expected: ${this.activeChain.id}, got: ${data['chain_id']}`);
+                }
+            }).catch(() => {
+                server.latency = -1;
+                resolve();
+            });
+        });
+    }
+
+    callStartupConn(server) {
+        if (this.connected === true) {
+            this.selectedEndpoint.next(server);
+            this.startup(null).catch(console.log);
+        }
+    }
+
+    async startup(url) {
+        if (this.isStarting) {
+            console.log('startup was already running...');
+            return;
+        } else {
+            this.isStarting = true;
+        }
+        let endpoint = url;
+        if (!url) {
+            endpoint = this.selectedEndpoint.getValue().url;
+            console.log('using saved endpoint:', endpoint);
+        } else {
+            this.status = '';
+            console.log('startup called on: ', url);
+        }
+
+        this.networkingReady.next(false);
+
+        // prevent double load after quick connection mode
+        if (endpoint !== this.lastEndpoint || this.autoMode === true) {
+
+            // define endpoint and initialize rpc
+            this.eosjs.initRPC(endpoint, this.activeChain.id, this.activeChain.endpoints);
+            this.lastEndpoint = endpoint;
+            this.autoMode = false;
+            this.defaultChains.find(c => c.id === this.activeChain.id).lastNode = this.lastEndpoint;
+
+            // get saved accounts
+            const savedAccounts = this.aService.readStoredAccounts();
+
+            if (savedAccounts && savedAccounts.length > 0) {
+
+                // load saved accounts for this chain
+                console.log(`loading local data for ${savedAccounts.length} accounts...`);
+                await this.aService.loadLocalAccounts(savedAccounts);
+
+                if (this.aService.lastAccount) {
+                    this.aService.select(this.aService.accounts.findIndex((a) => {
+                        return a.name === this.aService.lastAccount;
+                    }));
+                } else {
+                    this.aService.initFirst();
+                }
+
+                // navigate to the wallet page
+                await this.router['navigate'](['dashboard', 'wallet']);
+
+            } else {
+                if (this.crypto.getLockStatus()) {
+                    await this.router['navigate'](['']);
+                } else {
+                    // navigate to landing if on any other page
+                    if (this.router.url !== '/landing') {
+                        await this.router['navigate'](['landing']);
+                    }
+                }
+            }
+
+            this.networkingReady.next(true);
+
+            this.events.emit({event: 'connected'});
+
+            if (this.connectionTimeout) {
+                clearTimeout(this.connectionTimeout);
+                this.isStarting = false;
+                this.connectionTimeout = null;
+            }
+        }
+    }
 
 }
