@@ -3,7 +3,7 @@ import {Injectable, OnInit} from '@angular/core';
 import * as socketIo from 'socket.io-client';
 import {BehaviorSubject} from 'rxjs';
 import {AccountsService} from './accounts.service';
-import {EOSJSService} from './eosjs.service';
+import {Eosjs2Service} from './eosio/eosjs2.service';
 
 @Injectable({
 	providedIn: 'root'
@@ -20,9 +20,18 @@ export class RamService {
 	rm_base = 0;
 	rm_quote = 0;
 	rm_supply = 0;
-	reloaderInterval = null;
 
-	constructor(private aService: AccountsService, private eos: EOSJSService) {
+	restrictedChains = [
+		'EOS MAINNET',
+		'LIBERLAND TESTNET',
+		'LIBERLAND T2',
+		'LIBERLAND TEST LEGACY'
+	];
+
+	constructor(
+		private aService: AccountsService,
+		private eosjs: Eosjs2Service
+	) {
 		this.socket = socketIo('https://hapi.eosrio.io/');
 		this.socket.on('ticker', (data) => {
 			if (data.price) {
@@ -38,13 +47,13 @@ export class RamService {
 	}
 
 	reload() {
-		if (this.aService.activeChain.name !== 'EOS MAINNET') {
-			this.eos.getChainInfo().then((global) => {
+		if (!this.restrictedChains.includes(this.aService.activeChain.name)) {
+			this.eosjs.getChainInfo().then((global) => {
 				if (global) {
 					this.max_ram_size = global.rows[0]['max_ram_size'];
 					this.total_ram_bytes_reserved = global.rows[0]['total_ram_bytes_reserved'];
 					this.total_ram_stake = global.rows[0]['total_ram_stake'];
-					this.eos.getRamMarketInfo().then((rammarket) => {
+					this.eosjs.getRamMarketInfo().then((rammarket) => {
 						this.rm_base = rammarket.rows[0]['base']['balance'].split(' ')[0];
 						this.rm_quote = rammarket.rows[0]['quote']['balance'].split(' ')[0];
 						this.rm_supply = rammarket.rows[0]['supply'].split(' ')[0];
