@@ -1,442 +1,750 @@
 import {Injectable} from '@angular/core';
 
 export const DefaultTheme = {
-	'--main-background': '#272727',
-	'--background-gd': 'linear-gradient(to right, #232526, #414345)',
-	'--background-gd-webkit': '-webkit-linear-gradient(to right, #232526, #414345)',
-	'--header-background': '#313131',
-	'--header-title-color': '#9a9a9a',
-	'--header-icon-color': '#fafafa',
-	'--theme-icon-color': '#fafafa',
-	'--theme-icon-color-hover': '#c0c0c0',
+	/*dark theme colors*/
+	'--black-sidenav': '#191919',
+	'--black-background': '#272727',
+	'--black-card': '#333333',
+	'--dark-hover': '#1f1f1f', /*hover on dark btns, for example*/
+	'--blue-color': '#0094d2',
+	'--blue-darker': '#0077b7',
+	'--white-color': '#ffffff',
+	'--gray-lighter': '#e3e3e3', /*modal title, tab/sidenav text selected or hover, for example*/
+	'--gray-lightmedium': '#b1b1b1', /*input*/
+	'--gray-medium': '#a4a4a4', /*text card, text modal, for example*/
+	'--gray-darker': '#696969', /*subtitles, faded information*/
+	'--red-color': '#fb3c2c',
+
+	/*concepts*/
+	'--text-card-title-color': 'var(--gray-lighter)',
+	'--text-free-title-color': 'var(--gray-lighter)',
+	'--text-card-color': 'var(--gray-medium)',
+	'--text-highlight': 'var(--white-color)',
+	'--text-highlight-primary': 'var(--blue-color)',
+	'--text-highlight-accent': 'var(--blue-color)',
+	'--text-error': 'var(--red-color)',
+	'--text-faded': 'var(--gray-darker)',
+
+	/*buttons*/
+	/*solid*/
+	'--btn-primary-bg': 'var(--blue-color)',
+	'--btn-primary-hover': 'var(--blue-darker)',
+	'--btn-primary-color': 'var(--white-color)',
+	/*outline info*/
+	'--btn-info-outline': 'var(--blue-color)',
+	'--btn-info-outline-hover': 'var(--blue-color)',
+	'--btn-info-outline-color-hover': 'var(--white-color)',
+	/*outline inverse*/
+	'--btn-inverse-outline': 'var(--white-color)',
+	'--btn-inverse-outline-hover': 'var(--blue-color)',
+	'--btn-inverse-outline-color-hover': 'var(--white-color)',
+	/*outline accent*/
+	'--btn-accent-outline': 'var(--blue-color)',
+	'--btn-accent-outline-hover': 'var(--blue-color)',
+	'--btn-accent-outline-color-hover': 'var(--white-color)',
+	/*link*/
+	/*used in... refresh btn, open account on explorer, reload all proxys*/
+	'--btn-primary-link-color': 'var(--blue-color)',
+	'--btn-primary-link-color-hover': 'var(--blue-darker)',
+	'--btn-primary-link-bg-hover': 'var(--white-color)',
+	/*used in... retry connect, clear all data and logout, made with love by ->EOS Rio<-, view on github (on update)*/
+	'--btn-accent-link-color': 'var(--blue-color)',
+	'--btn-accent-link-color-hover': 'var(--blue-darker)',
+	'--btn-accent-link-bg-hover': 'var(--white-color)',
+
+	/*typography*/
 	'--special-font': 'inherit',
 	'--special-font-regular': 'inherit',
 	'--special-font-bold': 'inherit',
 	'--special-font-italic': 'inherit',
 	'--special-font-bolditalic': 'inherit',
-	'--slogan-special-color': 'inherit',
-	'--dashboard-animation-display': 'block',
-	'--dashboard-img-display': 'none',
 
+	/*main and body*/
+	'--main-background': 'var(--black-background)',
+	'--background-gd': 'linear-gradient(to right, #232526, #414345)',
+	'--background-gd-webkit': '-webkit-linear-gradient(to right, #232526, #414345)',
+	'--header-background': 'var(--black-background)',
+	'--header-title-color': 'var(--gray-medium)',
+	'--header-icon-color': 'var(--gray-lighter)',
+	'--theme-icon-color': 'var(--gray-lighter)',
+	'--theme-icon-color-hover': 'var(--gray-medium)',
+	'--error-bg': '#fbbdbd',
+
+	/*landing*/
 	'--landing-background': 'url("./assets/fundosimpleos.png")',
+	'--slogan-special-color': 'inherit',
 
-	'--sidenav-background': '#191919',
-	'--sidenav-color': '#b1b1b1',
-	'--sidenav-option-color': '#525a60',
-	'--sidenav-active-background': '#272727',
-	'--sidenav-active-color': '#e3e3e3',
-	'--sidenav-active-border-color1': '#272727',
-	'--sidenav-active-border-color2': '#535353',
+	/*sidenav*/
+	'--dashboard-animation-display': 'block', // for sidenav, landing and lockscreen logo
+	'--dashboard-img-display': 'none',  // for sidenav, landing and lockscreen logo
+	'--sidenav-background': 'var(--black-sidenav)',
+	'--sidenav-color': 'var(--gray-lightmedium)',
+	'--sidenav-active-background': 'var(--black-background)',
+	'--sidenav-active-color': 'var(--gray-lighter)',
+	'--sidenav-active-border-color1': 'var(--black-background)',
+	'--sidenav-active-border-color2': 'var(--gray-darker)',
 
-	'--top-btn-backgorund': '#1f1f1f',
-	'--top-btn-color': '#f8f8f8',
-
-	'--mat-tab-active': '#3b3b3b',
-	'--mat-tab-bar-active': '#0094d2',
+	/*toolbar*/
+	'--top-btn-background': 'var(--dark-hover)',
+	'--top-btn-color': 'var(--gray-lighter)',
+	/*tabs*/
+	'--mat-tab-active': 'var(--black-card)',
+	'--mat-tab-bar-active': 'var(--blue-color)',
 	'--mat-tab-bar-height': '2px',
 
-	'--tokens-panel-background': '#191919',
-	'--destaque-color-token': '#0094d2',
-	'--token-title-color': '#b1b1b1',
-	'--token-price-color': '#696969',
-	'--letter-divider': '#3f3f3f',
-	'--contact-hover-color': '#b1b1b1',
+	/*vote tabs*/
+	'--vote-tab-color': 'var(--gray-darker)',
+	'--vote-tab-active': 'var(--black-card)',
+	'--vote-tab-active-color': 'var(--gray-lighter)',
 
-	'--accordion-hover-background': '#0078a3',
-	'--accordion-color': '#cccccc',
-	'--accordion-a-color': '#cccccc',
+	/*tokens panel*/
+	'--tokens-panel-background': 'var(--black-sidenav)',
+	'--destaque-color-token': 'var(--blue-color)',
+	'--token-title-color': 'var(--gray-lightmedium)',
+	'--token-price-color': 'var(--gray-darker)',
+	'--tokens-panel-hover': 'var(--dark-hover)',
 
-	'--info-card-background': '#333333',
-	'--card-hover-backgorund': '#3b3b3b',
-	'--expand-background': '#1c1c1c',
+	/*contacts panel*/
+	'--letter-divider': 'var(--black-background)',
+	'--letter-color': 'var(--gray-lighter)',
+	'--contact-hover-color': 'var(--gray-lightmedium)',
+	'--newcontact-btn-bg': 'transparent',
+	'--newcontact-btn-color': 'var(--white-color)',
+	'--newcontact-btn-bordertop': 'var(--white-color)',
+	'--newcontact-btn-bg-hover': 'var(--blue-color)',
+	'--newcontact-btn-color-hover': 'var(--white-color)',
 
-	'--text-calendar': '#FFFFFF',
-	'--text-disabled-calendar': '#4d4d4d',
-	'--scroll-track-background': '#191919',
-	'--scroll-background': '#191919',
-	'--scroll-thumb-background': '#555555',
-
-	'--table-th-color': '#cccccc',
-	'--table-th-background': '#333333',
+	/*history table*/
+	'--table-th-color': 'var(--gray-lighter)',
+	'--table-th-background': 'var(--black-card)',
 	'--table-tr-background': 'transparent',
-	'--table-tr-color': '#b1b1b1',
-	'--table-td-color': '#b1b1b1',
-
-	'--verify-card-text-color': '#f8f8f8',
-	'--verify-card-backgound-color': 'rgba(162, 162, 162, 0.12)',
-
-	'--text-title-color': '#a4a4a4',
-	'--text-btn-color-hover': '#ffffff',
-	'--text-gray-color': '#b1b1b1',
-	'--text-gray2-color': '#e9e9e9',
-	'--text-gray3-color': '#d6d6d6',
-	'--text-white-color': '#ffffff',
-	'--text-destaque-color': '#0094d2',
-	'--text-blue-color': '#0094d2',
-	'--text-error': '#d04d42',
-	'--made-with-love': '#a4a4a4',
-
-	'--a-link-color': '#0094d2',
-
-	'--modal-background': '#191919',
-	'--modal-color': '#858585',
-	'--modal-title-color': '#e3e3e3',
-	'--modal-side-title-color': '#e3e3e3',
-	'--modal-side-color': '#c2c2c2',
-	'--modal-side-background': '#272727',
-	'--modal-step-background': '#535353',
-	'--modal-step-line-background': '#535353',
-	'--modal-step-off-background': '#828181',
-
+	'--table-tr-color': 'var(--gray-lightmedium)',
+	'--table-td-color': 'var(--gray-lightmedium)',
 	'--tr-even-backgorund': 'transparent',
 	'--tr-odd-backgorund': 'transparent',
+	'--expand-background': 'var(--dark-hover)',
 	'--tr-border-bottom-color': '1px solid #3f3f3f',
 
-	'--tr-resource-even-backgorund': '#2C2C2C',
-	'--tr-resource-odd-color': '#ececec',
+	/*cards*/
+	'--info-card-background': 'var(--black-card)',
+	'--info-card-bg-hover': '#3b3b3b',
+
+	/*accordion*/
+	'--accordion-header-color': 'var(--gray-lightmedium)',
+	'--accordion-content-color': 'var(--gray-lightmedium)',
+	'--accordion-header-color-hover': 'var(--gray-lighter)',
+
+	/*calendar*/
+	'--text-calendar': 'var(--white-color)',
+	'--text-disabled-calendar': 'var(--gray-darker)',
+
+	/*scrollbar*/
+	'--scroll-track-background': 'var(--black-sidenav)',
+	'--scroll-background': 'var(--black-sidenav)',
+	'--scroll-thumb-background': 'var(--gray-darker)',
+
+	/* modal / dialog / wizard */
+	'--modal-background': 'var(--black-sidenav)',
+	'--modal-color': 'var(--gray-medium)',
+	'--modal-title-color': 'var(--gray-lighter)',
+	'--modal-side-title-color': 'var(--gray-lighter)',
+	'--modal-side-color': 'var(--gray-lighter)',
+	'--modal-side-background': 'var(--black-background)',
+	'--modal-step-background': '#535353',
+	'--modal-step-line-background': '#535353',
+	'--modal-step-off-color': 'var(--gray-lightmedium)',
+
+	/*  input */
+	'--input-label': 'var(--gray-lightmedium)',
+
+	/* lent resources table */
+	'--tr-resource-even-backgorund': 'var(--black-background)',
+	'--tr-resource-odd-color': 'var(--gray-lighter)',
 	'--tr-resource-odd-backgorund': 'transparent',
-	'--tr-hover': '#1f1f1f',
+	'--tr-hover': 'var(--dark-hover)',
 
-	'--rex-balance-title-color': '#b1b1b1',
-	'--rex-balance-info-color': '#ffffff',
-	'--rex-btn-color': '#0094d2',
-	'--rex-border-color': '#0094d2',
+	/*REX page*/
+	'--rex-border-color': 'var(--black-background)',
 
-	'--btn-border-color': '#0094d2',
-	'--btn-color': '#0094d2',
-	'--btn-success-color': '#ffffff',
-	'--btn-color-hover': '#ffffff',
-	'--btn-background-hover': '#0077b7',
-	'--btn-rev-color-hover': '#0094d2',
+	'--made-with-love': 'var(--gray-medium)',
 
-	'--btn-primary-bg': '#0079b8',
-	'--btn-primary-hover': '#004d8a',
-
-	'--btn-inverse-color': '#ffffff',
-	'--btn-inverse-background-hover': '#0094d2',
-	'--btn-inverse-color-hover': '#ffffff',
-
-	'--btn-inverse2-color': '#ffffff',
-	'--btn-inverse2-background-hover': '#ffffff',
-	'--btn-inverse2-color-hover': '#0094d2',
-
-	'--btn-link2-color': '#0094d2',
-	'--btn-link2-background-hover': '#ffffff',
-	'--btn-link2-color-hover': '#0077b7',
-
-	'--btn-link-color': '#0094d2',
-	'--btn-link-background-hover': '#ffffff',
-	'--btn-link-color-hover': '#0077b7',
-
+	// '--update-box-background': '#1f1f1f',
 	'--update-box-background': '#1f1f1f',
-	'--chain-icon-bg': '#c5c5c5',
+	'--update-box-border': 'var(--black-card)',
+	'--chain-icon-bg': 'var(--gray-lighter)',
 };
 
 export const LiberlandTheme = {
-	'--main-background': '#EDEDEC',
-	'--background-gd': 'linear-gradient(to right, #EDEDEC, #F8F8F8)',
-	'--background-gd-webkit': '-webkit-linear-gradient(to right, #EDEDEC, #F8F8F8)',
-	'--header-background': '#D1D1D1',
-	'--header-title-color': '#292929',
-	'--header-icon-color': '#20202095',
-	'--theme-icon-color': '#20202095',
-	'--theme-icon-color-hover': '#292929',
+	/*light theme colors*/
+	'--light-bg': '#EDEDEC',
+	'--light-card': '#F8F8F8',
+	'--white-color': '#ffffff',
+	'--blue-color': '#0094d2',
+	'--blue-darker': '#0081bf',
+	'--liberland-blue': '#2a566f',
+	'--liberland-blue-darker': '#0f3a53',
+	'--yellow-color': '#FCD215',
+	'--yellow-darker': '#ebc115',
+	'--black-color': '#000000', /*modal, card titles*/
+	'--gray-dark': '#232323',
+	'--gray-darkmedium': '#404040', /*text card, text modal, for example*/
+	'--gray-medium': '#585858',
+	'--gray-lightmedium': '#adadad', /*subtitles, faded information*/
+	'--gray-light': '#D6D6D6',
+	'--red-color': '#ff4237',
+
+	/*concepts*/
+	'--text-card-title-color': 'var(--black-color)',
+	'--text-free-title-color': 'var(--gray-medium)',
+	'--text-card-color': 'var(--gray-dark)',
+	'--text-highlight': 'var(--black-color)',
+	'--text-highlight-primary': 'var(--blue-color)',
+	'--text-highlight-accent': 'var(--yellow-color)',
+	'--text-error': 'var(--red-color)',
+	'--text-faded': 'var(--gray-medium)',
+
+	/*buttons*/
+	/*solid*/
+	'--btn-primary-bg': 'var(--liberland-blue)',
+	'--btn-primary-hover': 'var(--liberland-blue-darker)',
+	'--btn-primary-color': 'var(--white-color)',
+	/*outline info*/
+	'--btn-info-outline': 'var(--liberland-blue)',
+	'--btn-info-outline-hover': 'var(--liberland-blue)',
+	'--btn-info-outline-color-hover': 'var(--white-color)',
+	/*outline inverse*/
+	'--btn-inverse-outline': 'var(--white-color)',
+	'--btn-inverse-outline-hover': 'var(--yellow-color)',
+	'--btn-inverse-outline-color-hover': 'var(--liberland-blue)',
+	/*outline accent*/
+	'--btn-accent-outline': 'var(--yellow-color)',
+	'--btn-accent-outline-hover': 'var(--yellow-color)',
+	'--btn-accent-outline-color-hover': 'var(--white-color)',
+	/*link*/
+	/*used in... refresh btn, open account on explorer, reload all proxys*/
+	'--btn-primary-link-color': 'var(--blue-color)',
+	'--btn-primary-link-color-hover': 'var(--blue-darker)',
+	'--btn-primary-link-bg-hover': 'var(--white-color)',
+	/*used in... retry connect, clear all data and logout, made with love by ->EOS Rio<-, view on github (on update)*/
+	'--btn-accent-link-color': 'var(--yellow-color)',
+	'--btn-accent-link-color-hover': 'var(--yellow-darker)',
+	'--btn-accent-link-bg-hover': 'var(--white-color)',
+
+	/*typography*/
 	'--special-font-regular': 'Lora',
 	'--special-font-bold': 'Lora-Bold',
 	'--special-font-italic': 'Lora-Italic',
 	'--special-font-bolditalic': 'Lora-BoldItalic',
 	'--special-font-weight': 'bold',
 	'--special-font-spacing': '1.3px',
+
+	/*main and body*/
+	'--main-background': 'var(--light-bg)',
+	'--background-gd': 'linear-gradient(to right, #EDEDEC, #F8F8F8)',
+	'--background-gd-webkit': '-webkit-linear-gradient(to right, #EDEDEC, #F8F8F8)',
+	'--header-background': 'var(--gray-light)',
+	'--header-title-color': 'var(--gray-dark)',
+	'--header-icon-color': 'var(--gray-dark)',
+	'--theme-icon-color': 'var(--gray-medium)',
+	'--theme-icon-color-hover': 'var(--gray-dark)',
+
+	/*landing*/
 	'--slogan-special-color': '#ffffff61',
+	'--landing-background': 'url("./assets/cover-liberland4.png")',
+	'--text-on-cover': 'var(--gray-light)',
 
-	'--landing-background': 'url("./assets/cover-liberland3.png")',
 
-	'--sidenav-background': '#1577C6',
+	/*sidenav*/
+	'--dashboard-animation-display': 'block', // for sidenav, landing and lockscreen logo
+	'--dashboard-img-display': 'none',  // for sidenav, landing and lockscreen logo
+	'--sidenav-background': 'var(--liberland-blue)',
 	'--sidenav-color': '#D6D6D6',
-	'--sidenav-option-color': '#f8f8f8',
-	'--sidenav-active-background': '#EDEDEC',
-	'--sidenav-active-color': '#1577C6',
-	'--sidenav-active-border-color1': '#EDEDEC',
+	'--sidenav-active-background': 'var(--light-bg)',
+	'--sidenav-active-color': 'var(--liberland-blue)',
+	'--sidenav-active-border-color1': 'var(--light-bg)',
+	'--sidenav-active-border-color2': 'var(var(--liberland-blue-darker))',
 
-	'--top-btn-backgorund': '#084577',
-	'--top-btn-color': '#f8f8f8',
-
-	'--mat-tab-active': '#084577',
-	'--mat-tab-bar-active': '#FCD215',
+	/*toolbar*/
+	'--top-btn-background': 'var(--liberland-blue-darker)',
+	'--top-btn-color': 'var(--white-color)',
+	'--mat-tab-active': 'var(--liberland-blue-darker)',
+	'--mat-tab-bar-active': 'var(--yellow-color)',
 	'--mat-tab-bar-height': '4px',
 
-	'--tokens-panel-background': '#1577C6',
-	'--destaque-color-token': '#FCD215',
-	'--token-title-color': '#f8f8f8',
-	'--token-price-color': '#D6D6D6',
-	'--letter-divider': '#FCD215',
-	'--contact-hover-color': '#292929',
+	/*vote tabs*/
+	'--vote-tab-color': 'var(--gray-lightmedium)',
+	'--vote-tab-active': 'var(--light-card)',
+	'--vote-tab-active-color': 'var(--liberland-blue)',
 
-	'--accordion-hover-background': '#0078a3',
-	'--accordion-color': '#707070',
-	'--accordion-a-color': '#707070',
-	'--accordion-color-hover': '#404040',
+	/*tokens panel*/
+	'--tokens-panel-background': 'var(--liberland-blue)',
+	'--destaque-color-token': 'var(--yellow-color)',
+	'--token-title-color': 'var(--light-card)',
+	'--token-price-color': 'var(--gray-light)',
+	'--tokens-panel-hover': 'var(--liberland-blue-darker)',
 
-	'--info-card-background': '#F8F8F8',
-	'--card-hover-backgorund': '#ececec',
-	'--expand-background': '#FAFAFA',
+	/*contacts panel*/
+	'--letter-divider': 'var(--liberland-blue-darker)',
+	'--letter-color': 'var(--white-color)',
+	'--contact-hover-color': 'var(--liberland-blue-darker)',
+	'--newcontact-btn-bg': 'var(--liberland-blue-darker)',
+	'--newcontact-btn-color': 'var(--white-color)',
+	'--newcontact-btn-bordertop': 'var(--white-color)',
+	'--newcontact-btn-bg-hover': 'var(--yellow-color)',
+	'--newcontact-btn-color-hover': 'var(--liberland-blue-darker)',
 
-	'--text-calendar': '#404040',
-	'--text-disabled-calendar': '#d6d6d6',
+	/*history table*/
+	'--table-th-color': 'var(--liberland-blue)',
+	'--table-th-background': 'var(--light-card)',
+	'--table-tr-background': 'var(--white-color)',
+	'--table-tr-color': 'var(--gray-darkmedium)',
+	'--table-td-color': 'var(--gray-darkmedium)',
+	'--tr-even-backgorund': 'var(--white-color)',
+	'--tr-odd-backgorund': 'var(--white-color)',
+	'--expand-background': 'var(--light-card)',
+	'--tr-border-bottom-color': '1px solid var(--gray-light)',
+	'--tr-hover': 'var(--light-bg)',
 
+	/*cards*/
+	'--info-card-background': 'var(--light-card)',
+	'--info-card-bg-hover': 'var(--gray-light)',
+
+	/*accordion*/
+	'--accordion-header-color': 'var(--gray-darkmedium)',
+	'--accordion-content-color': 'var(--black-color)',
+	'--accordion-header-color-hover': 'var(--white-color)',
+
+	/*calendar*/
+	'--text-calendar': 'var(--gray-darkmedium)',
+	'--text-disabled-calendar': 'var(--gray-light)',
+
+	/*scrollbar*/
 	'--scroll-track-background': '#f1f1f1',
 	'--scroll-background': '#f1f1f1',
-	'--scroll-thumb-background': '#D6D6D6',
+	'--scroll-thumb-background': 'var(--gray-light)',
 
-	'--table-th-color': '#084577',
-	'--table-th-background': '#f8f8f8',
-	'--table-tr-background': '#ffffff',
-	'--table-tr-color': '#404040',
-	'--table-td-color': '#404040',
-
-	'--verify-card-text-color': '#404040',
-	'--verify-card-backgound-color': '#f8f8f8',
-
-	'--text-title-color': '#404040',
-	'--text-white-color': '#202020',
-	'--text-btn-color-hover': '#f8f8f8',
-	'--text-gray-color': '#404040',
-	'--text-gray2-color': '#707070',
-	'--text-gray3-color': '#b1b1b1',
-	'--text-destaque-color': '#FCD215',
-	'--text-blue-color': '#1577C6',
-	'--text-error': '#ff8880',
-	'--made-with-love': '#ffffff',
-
-	'--a-link-color': '#F8F8F8',
-
-	'--modal-background': '#F8F8F8',
+	/* modal / dialog / wizard */
+	'--modal-background': 'var(--light-card)',
 	'--modal-color': '#707070',
-	'--modal-title-color': '#404040',
-	'--modal-side-title-color': '#F8F8F8',
-	'--modal-side-color': '#d6d6d6',
-	'--modal-side-background': '#1577C6',
-	'--modal-step-background': '#084577',
-	'--modal-step-line-background': '#084577',
-	'--modal-step-off-background': '#ffffff',
+	'--modal-title-color': 'var(--gray-darkmedium)',
+	'--modal-side-title-color': 'var(--light-card)',
+	'--modal-side-color': 'var(--gray-light)',
+	'--modal-side-background': 'var(--liberland-blue)',
+	'--modal-step-background': 'var(--liberland-blue-darker)',
+	'--modal-step-line-background': 'var(--liberland-blue-darker)',
+	'--modal-step-off-color': 'var(--white-color)',
+	'--error-bg': '#f7e2e2',
 
-	'--tr-even-backgorund': '#ffffff',
-	'--tr-odd-backgorund': '#ffffff',
-	'--tr-border-bottom-color': '1px solid #cccccc',
+	/*  input */
+	'--input-label': 'var(--gray-darkmedium)',
 
-	'--tr-resource-even-backgorund': '#1577C6',
+	/* lent resources table */
+	'--tr-resource-even-backgorund': 'var(--white-color)',
 	'--tr-resource-odd-color': '#202020',
 	'--tr-resource-odd-backgorund': 'transparent',
-	'--tr-hover': '#d2d2d2',
 
-	'--rex-balance-title-color': '#cccccc',
-	'--rex-balance-info-color': '#f8f8f8',
-	'--rex-btn-color': '#ffffff',
-	'--rex-border-color': '#ffffff',
+	/*REX page*/
+	'--rex-border-color': 'var(--light-bg)',
 
-	'--btn-border-color': '#084577',
-	'--btn-color': '#084577',
-	'--btn-success-color': '#303030',
-	'--btn-color-hover': '#ffffff',
-	'--btn-background-hover': '#084577',
-	'--btn-rev-color-hover': '#1577C6',
-
-	'--btn-primary-bg': '#0079b8',
-	'--btn-primary-hover': '#004d8a',
-
-	'--btn-inverse-color': '#ffffff',
-	'--btn-inverse-background-hover': '#ffffff',
-	'--btn-inverse-color-hover': '#084577',
-
-	'--btn-inverse2-color': '#202020',
-	'--btn-inverse2-background-hover': '#ffffff',
-	'--btn-inverse2-color-hover': '#084577',
-
-	'--btn-link2-color': '#FCD215',
-	'--btn-link2-background-hover': '#ffffff',
-	'--btn-link2-color-hover': '#c39d13',
-
-	'--btn-link-color': '#fafafa',
-	// '--btn-link-color': '#1577C6',
-	'--btn-link-background-hover': '#ffffff',
-	'--btn-link-color-hover': '#084577',
-
-	'--update-box-background': '#0d3352',
+	'--update-box-background': 'var(--light-card)',
+	'--update-box-border': 'var(--liberland-blue-darker)',
 	'--chain-icon-bg': '#c5c5c5',
 };
 
 export const LightTheme = {
-	'--main-background': '#ffffff',
-	'--background-gd': 'linear-gradient(to right, #EDEDEC, #F8F8F8)',
-	'--background-gd-webkit': '-webkit-linear-gradient(to right, #EDEDEC, #F8F8F8)',
-	'--header-background': '#D1D1D1',
-	'--header-title-color': '#292929',
-	'--header-icon-color': '#20202095',
-	'--theme-icon-color': '#20202095',
-	'--theme-icon-color-hover': '#292929',
+	/*light theme colors*/
+	'--white-color': '#ffffff',
+	'--light-card': '#f9f9f9',
+	'--blue-color': '#0094d2',
+	'--blue-darker': '#0081bf',
+	'--yellow-color': '#FCD215',
+	'--black-color': '#000000', /*modal, card titles*/
+	'--gray-dark': '#232323',
+	'--gray-darkmedium': '#404040', /*text card, text modal, for example*/
+	'--gray-medium': '#585858',
+	'--gray-lightmedium': '#adadad', /*subtitles, faded information*/
+	'--gray-light': '#dadada',
+	'--red-color': '#f44336',
+
+	/*concepts*/
+	'--text-card-title-color': 'var(--black-color)',
+	'--text-free-title-color': 'var(--gray-lightmedium)',
+	'--text-card-color': 'var(--gray-dark)',
+	'--text-highlight': 'var(--black-color)',
+	'--text-highlight-primary': 'var(--blue-color)',
+	'--text-highlight-accent': 'var(--yellow-color)',
+	'--text-error': 'var(--red-color)',
+	'--text-faded': 'var(--gray-medium)',
+
+	/*buttons*/
+	/*solid*/
+	'--btn-primary-bg': 'var(--blue-color)',
+	'--btn-primary-hover': 'var(--blue-darker)',
+	'--btn-primary-color': 'var(--white-color)',
+	/*outline info*/
+	'--btn-info-outline': 'var(--blue-color)',
+	'--btn-info-outline-hover': 'var(--blue-color)',
+	'--btn-info-outline-color-hover': 'var(--white-color)',
+	/*outline inverse*/
+	'--btn-inverse-outline': 'var(--gray-dark)',
+	'--btn-inverse-outline-hover': 'var(--blue-color)',
+	'--btn-inverse-outline-color-hover': 'var(--white-color)',
+	/*outline accent*/
+	'--btn-accent-outline': 'var(--blue-color)',
+	'--btn-accent-outline-hover': 'var(--blue-color)',
+	'--btn-accent-outline-color-hover': 'var(--white-color)',
+	/*link*/
+	/*used in... refresh btn, open account on explorer, reload all proxys*/
+	'--btn-primary-link-color': 'var(--blue-color)',
+	'--btn-primary-link-color-hover': 'var(--blue-darker)',
+	'--btn-primary-link-bg-hover': 'var(--white-color)',
+	/*used in... retry connect, clear all data and logout, made with love by ->EOS Rio<-, view on github (on update)*/
+	'--btn-accent-link-color': 'var(--blue-color)',
+	'--btn-accent-link-color-hover': 'var(--blue-darker)',
+	'--btn-accent-link-bg-hover': 'var(--white-color)',
+
+	/*typography*/
 	'--special-font': 'inherit',
 	'--special-font-regular': 'inherit',
 	'--special-font-bold': 'inherit',
 	'--special-font-italic': 'inherit',
 	'--special-font-bolditalic': 'inherit',
 	'--special-font-weight': 'inherit',
-	'--slogan-special-color': '#adadad',
 
+	/*main and body*/
+	'--main-background': 'var(--white-color)',
+	'--background-gd': 'linear-gradient(to right, #EDEDEC, #F8F8F8)',
+	'--background-gd-webkit': '-webkit-linear-gradient(to right, #EDEDEC, #F8F8F8)',
+	'--header-background': 'var(--gray-light)',
+	'--header-title-color': 'var(--gray-dark)',
+	'--header-icon-color': 'var(--gray-dark)',
+	'--theme-icon-color': 'var(--gray-medium)',
+	'--theme-icon-color-hover': 'var(--gray-dark)',
+
+	/*landing*/
+	'--slogan-special-color': 'var(--gray-medium)',
 	'--landing-background': 'url("./assets/fundosimpleoslight.png")',
 	// '--landing-image-background': 'url("../../assets/fundosimpleoslight.png")',
+	'--text-on-cover': 'var(--gray-medium)',
 
+	/*sidenav*/
+	'--dashboard-animation-display': 'none', // for sidenav, landing and lockscreen logo
+	'--dashboard-img-display': 'block', // for sidenav, landing and lockscreen logo
+	'--sidenav-background': 'var(--blue-color)',
+	'--sidenav-color': 'var(--white-color)',
+	'--sidenav-active-background': 'var(--white-color)',
+	'--sidenav-active-color': 'var(--blue-color)',
+	'--sidenav-active-border-color1': 'var(--white-color)',
+	'--sidenav-active-border-color2': 'var(--gray-dark)',
 
-	'--dashboard-animation-display': 'none',
-	'--dashboard-img-display': 'block',
-
-	// '--sidenav-background': '#f8f8f8',
-	'--sidenav-background': '#0094d2',
-	// '--sidenav-color': '#565656',
-	'--sidenav-color': '#dadada',
-	// '--sidenav-option-color': '#a5a5a5',
-	'--sidenav-option-color': '#f8f8f8',
-	'--sidenav-active-background': '#ffffff',
-	// '--sidenav-active-color': '#9a9a9a',
-	'--sidenav-active-color': '#0094d2',
-	// '--sidenav-active-border-color1': '#ffffff',
-	'--sidenav-active-border-color1': '#EDEDEC',
-
-	'--top-btn-backgorund': '#afafaf',
-	'--top-btn-color': '#f8f8f8',
-
-	'--mat-tab-active': '#afafaf',
-	'--mat-tab-bar-active': '#FCD215',
+	/*toolbar*/
+	'--top-btn-background': 'var(--blue-darker)',
+	'--top-btn-color': 'var(--white-color)',
+	'--mat-tab-active': 'var(--blue-darker)',
+	'--mat-tab-bar-active': 'var(--yellow-color)',
 	'--mat-tab-bar-height': '4px',
 
-	'--tokens-panel-background': '#f8f8f8',
-	'--destaque-color-token': '#0094d2',
-	'--token-title-color': '#000000',
-	'--token-price-color': '#707070',
-	'--letter-divider': '#afafaf',
-	'--contact-hover-color': '#ffffff',
+	/*vote tabs*/
+	'--vote-tab-color': 'var(--gray-lightmedium)',
+	'--vote-tab-active': 'var(--light-card)',
+	'--vote-tab-active-color': 'var(--blue-color)',
 
-	'--accordion-hover-background': '#bfbfbf',
-	'--accordion-color': '#707070',
-	'--accordion-a-color': '#707070',
-	'--accordion-color-hover': '#404040',
+	/*tokens panel*/
+	'--tokens-panel-background': 'var(--light-card)',
+	'--destaque-color-token': 'var(--blue-color)',
+	'--token-title-color': 'var(--black-color)',
+	'--token-price-color': 'var(--gray-medium)',
+	'--tokens-panel-hover': 'var(--gray-light)',
 
-	'--info-card-background': '#F8F8F8',
-	'--card-hover-backgorund': '#ececec',
-	'--expand-background': '#FAFAFA',
+	/*contacts panel*/
+	'--letter-divider': 'var(--gray-lightmedium)',
+	'--letter-color': 'var(--black-color)',
+	'--contact-hover-color': 'var(--gray-light)',
+	'--newcontact-btn-bg': 'var(--gray-light)',
+	'--newcontact-btn-color': 'var(--blue-color)',
+	'--newcontact-btn-bordertop': 'var(--blue-color)',
+	'--newcontact-btn-bg-hover': 'var(--blue-color)',
+	'--newcontact-btn-color-hover': 'var(--white-color)',
 
-	'--text-calendar': '#404040',
-	'--text-disabled-calendar': '#d6d6d6',
+	/*history table*/
+	'--table-th-color': 'var(--gray-dark)',
+	'--table-th-background': 'var(--light-card)',
+	'--table-tr-background': 'var(--white-color)',
+	'--table-tr-color': 'var(--gray-darkmedium)',
+	'--table-td-color': 'var(--gray-darkmedium)',
+	'--tr-even-backgorund': 'var(--white-color)',
+	'--tr-odd-backgorund': 'var(--white-color)',
+	'--expand-background': 'var(--light-card)',
+	'--tr-border-bottom-color': '1px solid var(--gray-light)',
+	'--tr-hover': 'var(--gray-light)',
 
-	'--scroll-track-background': '#f1f1f1',
-	'--scroll-background': '#f1f1f1',
-	'--scroll-thumb-background': '#D6D6D6',
+	/*cards*/
+	'--info-card-background': 'var(--light-card)',
+	'--info-card-bg-hover': 'var(--gray-light)',
 
-	'--table-th-color': '#212121',
-	'--table-th-background': '#f8f8f8',
-	'--table-tr-background': '#ffffff',
-	'--table-tr-color': '#404040',
-	'--table-td-color': '#404040',
+	/*accordion*/
+	'--accordion-header-color': 'var(--gray-darkmedium)',
+	'--accordion-content-color': 'var(--black-color)',
+	'--accordion-header-color-hover': 'var(--white-color)',
 
-	'--verify-card-text-color': '#696969',
-	'--text-title-color': '#404040',
-	'--text-white-color': '#202020',
-	'--text-btn-color-hover': '#f8f8f8',
-	'--text-gray-color': '#404040',
-	'--text-gray2-color': '#707070',
-	'--text-gray3-color': '#b1b1b1',
-	'--text-destaque-color': '#FCD215',
-	'--text-blue-color': '#0094d2',
-	'--text-error': '#f44336',
-	'--made-with-love': '#404040',
+	/*calendar*/
+	'--text-calendar': 'var(--gray-darkmedium)',
+	'--text-disabled-calendar': 'var(--gray-light)',
 
-	'--a-link-color': '#FCD215',
+	/*scrollbar*/
+	'--scroll-track-background': 'var(--light-card)',
+	'--scroll-background': 'var(--light-card)',
+	'--scroll-thumb-background': 'var(--gray-light)',
 
-	'--modal-background': '#F8F8F8',
+	/* modal / dialog / wizard */
+	'--modal-background': 'var(--light-card)',
 	'--modal-color': '#707070',
 	'--modal-title-color': '#404040',
 	'--modal-side-title-color': '#F8F8F8',
 	'--modal-side-color': '#d6d6d6',
-	'--modal-side-background': '#0094d2',
-	'--modal-step-background': '#afafaf',
-	'--modal-step-line-background': '#afafaf',
-	'--modal-step-off-background': '#ffffff',
+	'--modal-side-background': 'var(--blue-color)',
+	'--modal-step-background': 'var(--blue-darker)',
+	'--modal-step-line-background': 'var(--blue-darker)',
+	'--modal-step-off-color': 'var(--white-color)',
+	'--error-bg': '#f7e2e2',
 
-	'--tr-even-backgorund': '#ffffff',
-	'--tr-odd-backgorund': '#ffffff',
-	'--tr-border-bottom-color': '1px solid #cccccc',
-	'--tr-resource-even-backgorund': '#aaaaaa',
+	/*  input */
+	'--input-label': 'var(--gray-darkmedium)',
+
+	/* lent resources table */
+	'--tr-resource-even-backgorund': 'var(--white-color)',
 	'--tr-resource-odd-color': '#202020',
-	'--tr-resource-odd-backgorund': 'transparent',
-	'--tr-hover': '#afafaf',
+	'--tr-resource-odd-backgorund': 'var(--light-card)',
 
-	'--rex-balance-title-color': '#e9e9e9',
-	'--rex-balance-info-color': '#f8f8f8',
-	'--rex-btn-color': '#ffffff',
-	'--rex-border-color': '#ffffff',
+	/*REX page*/
+	'--rex-border-color': 'var(--gray-lightmedium)',
 
-	'--btn-border-color': '#0094d2',
-	'--btn-color': '#0094d2',
-	'--btn-success-color': '#303030',
-	'--btn-color-hover': '#ffffff',
-	'--btn-background-hover': '#0094d2',
-	'--btn-rev-color-hover': '#0094d2',
+	'--made-with-love': '#404040',
 
-	'--btn-primary-bg': '#0079b8',
-	'--btn-primary-hover': '#004d8a',
-
-	'--btn-inverse-color': '#0094d2',
-	'--btn-inverse-background-hover': '#0094d2',
-	'--btn-inverse-color-hover': '#ffffff',
-
-	'--btn-inverse2-color': '#404040',
-	'--btn-inverse2-background-hover': '#ffffff',
-	'--btn-inverse2-color-hover': '#0094d2',
-
-	'--btn-link2-color': '#0079b8',
-	'--btn-link2-background-hover': '#ffffff',
-	'--btn-link2-color-hover': '#0066a6',
-
-	'--btn-link-color': '#0094d2',
-	'--btn-link-background-hover': '#ffffff',
-	'--btn-link-color-hover': '#0077b7',
-
-	'--update-box-background': '#a4a4a4',
+	'--update-box-background': 'var(--white-color)',
+	'--update-box-border': 'var(--gray-lightmedium)',
 	'--chain-icon-bg': '#c5c5c5',
 
 };
 
 export const WaxLandingTheme = {
+	/*wax theme colors*/
+	'--white-color': '#ffffff',
+	'--blue-card': '#002731',
+	'--blue-card-darker': '#001d24',
+	'--blue-color': '#0094d2',
+	'--blue-darker': '#0081bf',
+	'--orange-color': '#F78E1E',
+	'--orange-darker': '#e8964a',
+	'--black-color': '#000000', /*modal, card titles*/
+	'--gray-dark': '#232323',
+	'--gray-darkmedium': '#404040', /*text card, text modal, for example*/
+	'--gray-medium': '#585858',
+	'--gray-lightmedium': '#c0c0c0', /*subtitles, faded information*/
+	'--gray-light': '#e3e3e3',
+	'--red-color': '#ff5346',
+
+	/*concepts*/
+	'--text-card-title-color': 'var(--white-color)',
+	// '--text-free-title-color': 'var(--gray-lightmedium)',
+	'--text-card-color': 'var(--gray-light)',
+	'--text-highlight': 'var(--white-color)',
+	'--text-highlight-primary': 'var(--orange-color)',
+	'--text-highlight-accent': 'var(--orange-color)',
+	'--text-error': 'var(--red-color)',
+	'--text-faded': 'var(--gray-lightmedium)',
+
+	/*buttons*/
+	/*solid*/
+	'--btn-primary-bg': 'var(--orange-color)',
+	'--btn-primary-hover': 'var(--orange-color)',
+	'--btn-primary-color': 'var(--white-color)',
+	/*outline info*/
+	'--btn-info-outline': 'var(--orange-color)',
+	'--btn-info-outline-hover': 'var(--orange-color)',
+	'--btn-info-outline-color-hover': 'var(--white-color)',
+	/*outline inverse*/
+	'--btn-inverse-outline': 'var(--white-color)',
+	'--btn-inverse-outline-hover': 'var(--orange-color)',
+	'--btn-inverse-outline-color-hover': 'var(--white-color)',
+	/*outline accent*/
+	// '--btn-accent-outline': 'var(--blue-color)',
+	// '--btn-accent-outline-hover': 'var(--blue-color)',
+	// '--btn-accent-outline-color-hover': 'var(--white-color)',
+	/*link*/
+	/*used in... refresh btn, open account on explorer, reload all proxys, view on github (on update)*/
+	'--btn-primary-link-color': 'var(--orange-color)',
+	'--btn-primary-link-color-hover': 'var(--orange-color)',
+	/*used in... retry connect, clear all data and logout, made with love by ->EOS Rio<-, */
+	'--btn-accent-link-color': 'var(--orange-color)',
+	'--btn-accent-link-color-hover': 'var(--orange-darker)',
+
+	/*main and body*/
+	'--header-background': 'var(--gray-darkmedium)',
+	'--header-title-color': 'var(--gray-light)',
+	'--header-icon-color': 'var(---gray-light)',
+	'--theme-icon-color': 'var(--gray-light)',
+	'--theme-icon-color-hover': 'var(--white-color)',
+
+	/*landing*/
+	'--slogan-special-color': 'var(--gray-lightmedium)',
 	'--landing-background': 'url("./assets/cover-wax.jpg")',
-	'--info-card-background': '#002731',
-	'--btn-inverse-background-hover': '#F78E1E',
+	'--text-on-cover': 'var(--gray-lightmedium)',
+	'--dashboard-animation-display': 'block', // for sidenav, landing and lockscreen logo
+	'--dashboard-img-display': 'none',  // for sidenav, landing and lockscreen logo
+
+	/*cards*/
+	'--info-card-background': 'var(--blue-card)',
+	'--info-card-bg-hover': 'var(--blue-card-darker)',
+
+	/*scrollbar*/
+	'--scroll-track-background': 'var(--blue-card-darker)',
+	'--scroll-background': '(var(--blue-card-darker)',
+	'--scroll-thumb-background': 'var(--gray-darkmedium)',
+
+	/* modal / dialog / wizard */
+	'--modal-background': 'var(--blue-card)',
+	'--modal-color': 'var(--gray-light)',
+	'--modal-title-color': 'var(--white-color)',
+	'--modal-side-title-color': 'var(--white-color)',
+	'--modal-side-color': 'var(--gray-light)',
+	// '--modal-side-color': '#d6d6d6',
+	'--modal-side-background': 'var(--blue-card-darker)',
+	'--modal-step-background': 'var(--orange-color)',
+	'--modal-step-line-background': 'var(--orange-color)',
+	'--modal-step-off-color': 'var(--white-color)',
+
+	/*  input */
+	'--input-label': 'var(--gray-light)',
+
+	/*calendar*/
+	'--text-calendar': 'var(--white-color)',
+
 	'--text-title-color': '#d0d0d0',
-	'--text-blue-color': '#F78E1E',
-	'--update-box-background': '#002731',
-	'--btn-link-color': '#F78E1E',
-	'--btn-link-color-hover': '#e8964a',
-	'--btn-link2-color': '#F78E1E',
-	'--btn-rev-color-hover': '#F78E1E',
-	'--modal-background': '#002731',
-	'--btn-border-color': '#F78E1E',
-	'--btn-color': '#F78E1E',
-	'--btn-background-hover': '#F78E1E',
-	'--modal-side-background': '#001d24',
-	'--modal-step-off-background': '#ffffff',
-	'--btn-primary-bg': '#c56907',
-	'--btn-primary-hover': '#8f4c05',
-	'--chain-icon-bg': 'white',
-	'--text-error': '#ff8e85',
+
+	'--chain-icon-bg': 'var(--white-color)',
 	'--made-with-love': '#bfbfbf',
-	'--modal-title-color': '#ffffff',
+
+	'--update-box-background': 'var(--blue-card)',
+};
+
+export const TelosLandingTheme = {
+	/*telos theme colors*/
+	'--white-color': '#ffffff',
+	'--light-card': '#f9f9f9',
+	'--purple-bg': '#030039',
+	'--blue-card-darker': '#001d24',
+	'--blue-color': '#0094d2',
+	'--blue-darker': '#0081bf',
+	'--telos-purple': '#571AFF',
+	'--telos-purple-darker': '#3D12B2',
+	'--yellow-color':'#FFD75E',
+	'--yellow-darker':'#eec65a',
+	'--black-color': '#000000', /*modal, card titles*/
+	'--gray-dark': '#232323',
+	'--gray-darkmedium': '#404040', /*text card, text modal, for example*/
+	'--gray-medium': '#585858',
+	'--gray-lightmedium': '#c0c0c0', /*subtitles, faded information*/
+	'--gray-light': '#e3e3e3',
+	'--red-color': '#ff1301',
+
+	/*concepts*/
+	'--text-card-title-color': 'var(--black-color)',
+	// '--text-free-title-color': 'var(--gray-lightmedium)',
+	'--text-card-color': 'var(--gray-dark)',
+	'--text-highlight': 'var(--black-color)',
+	'--text-highlight-primary': 'var(--telos-purple)',
+	'--text-highlight-accent': 'var(--yellow-color)',
+	'--text-error': 'var(--red-color)',
+	'--text-faded': 'var(--gray-medium)',
+
+	/*buttons*/
+	/*solid*/
+	'--btn-primary-bg': 'var(--telos-purple)',
+	'--btn-primary-hover': 'var(--telos-purple-darker)',
+	'--btn-primary-color': 'var(--white-color)',
+	/*outline info*/
+	'--btn-info-outline': 'var(--telos-purple)',
+	'--btn-info-outline-hover': 'var(--telos-purple)',
+	'--btn-info-outline-color-hover': 'var(--white-color)',
+	/*outline inverse*/
+	'--btn-inverse-outline': 'var(--white-color)',
+	'--btn-inverse-outline-hover': 'var(--telos-purple)',
+	'--btn-inverse-outline-color-hover': 'var(--white-color)',
+	/*outline accent*/
+	// '--btn-accent-outline': 'var(--blue-color)',
+	// '--btn-accent-outline-hover': 'var(--blue-color)',
+	// '--btn-accent-outline-color-hover': 'var(--white-color)',
+	/*link*/
+	/*used in... refresh btn, open account on explorer, reload all proxys, view on github (on update)*/
+	'--btn-primary-link-color': 'var(--telos-purple)',
+	'--btn-primary-link-color-hover': 'var(--telos-purple-darker)',
+	/*used in... retry connect, clear all data and logout, made with love by ->EOS Rio<-, */
+	'--btn-accent-link-color': 'var(--yellow-color)',
+	'--btn-accent-link-color-hover': 'var(--telos-purple)',
+
+	/*main and body*/
+	'--header-background': 'var(--gray-darkmedium)',
+	'--header-title-color': 'var(--gray-light)',
+	'--header-icon-color': 'var(---gray-light)',
+	'--theme-icon-color': 'var(--gray-light)',
+	'--theme-icon-color-hover': 'var(--white-color)',
+
+	/*landing*/
+	'--slogan-special-color': 'var(--gray-lightmedium)',
+	'--landing-background': 'url("./assets/cover-telos.jpg")',
+	'--text-on-cover': 'var(--gray-lightmedium)',
+	'--dashboard-animation-display': 'block', // for sidenav, landing and lockscreen logo
+	'--dashboard-img-display': 'none',  // for sidenav, landing and lockscreen logo
+
+	/*cards*/
+	'--info-card-background': 'var(--light-card)',
+	'--info-card-bg-hover': 'var(--gray-light)',
+
+	/*calendar*/
+	'--text-calendar': 'var(--gray-medium)',
+
+	/*scrollbar*/
+	'--scroll-track-background': 'var(--gray-light)',
+	'--scroll-background': 'var(--gray-light)',
+	'--scroll-thumb-background': 'var(--gray-lightmedium)',
+
+	/* modal / dialog / wizard */
+	'--modal-background': 'var(--light-card)',
+	'--modal-color': 'var(--gray-dark)',
+	'--modal-title-color': 'var(--black-color)',
+	'--modal-side-title-color': 'var(--white-color)',
+	'--modal-side-color': 'var(--gray-light)',
+	'--modal-side-background': 'var(--purple-bg)',
+	'--modal-step-background': 'var(--telos-purple)',
+	'--modal-step-line-background': 'var(-telos-purple)',
+	'--modal-step-off-color': 'var(--white-color)',
+
+	/*  input */
+	'--input-label': 'var(--gray-darkmedium)',
+
+	'--text-title-color': '#d0d0d0',
+
+	'--chain-icon-bg': 'var(--white-color)',
+	'--made-with-love': '#bfbfbf',
+
+	'--update-box-background': 'var(--light-card)',
 };
 
 @Injectable({providedIn: 'root'})
 export class ThemeService {
+
+    lightMode = false;
 	public currentTheme: string;
 
 	defaultTheme() {
 		this.currentTheme = 'dark';
 		this.setTheme(DefaultTheme);
+	}
+
+	lightTheme() {
+		this.currentTheme = 'light';
+		this.setTheme(LightTheme);
 	}
 
 	liberlandTheme() {
@@ -449,9 +757,9 @@ export class ThemeService {
 		this.setTheme(WaxLandingTheme);
 	}
 
-	lightTheme() {
-		this.currentTheme = 'light';
-		this.setTheme(LightTheme);
+	telosTheme() {
+		this.currentTheme = 'telos';
+		this.setTheme(TelosLandingTheme);
 	}
 
 	private setTheme(theme: {}) {
