@@ -3,7 +3,7 @@ import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 
 import {AppModule} from './app/app.module';
 import {environment} from './environments/environment';
-import 'isomorphic-fetch';
+import 'node-fetch';
 
 import 'hammerjs';
 
@@ -11,23 +11,48 @@ import 'echarts/theme/macarons.js';
 import 'echarts/map/js/world.js';
 import 'echarts/dist/extension/bmap.min.js';
 
-if (environment.production) {
-	enableProdMode();
+function angularBoot() {
+    if (environment.production) {
+        enableProdMode();
+    }
+    platformBrowserDynamic().bootstrapModule(AppModule, {
+        preserveWhitespaces: false
+    }).catch(err => console.log(err));
 }
 
-const url = 'https://raw.githubusercontent.com/eosrio/simpleos/master/config.json';
-fetch( url ).then(function(response) {
-	if (response.status >= 400) {
-		throw new Error("Bad response from server");
-	}
-	return response.json();
-}).then(function(result) {
-		// console.log(result);
-	const payload = {lastUpdate: new Date(), config: result };
-	localStorage.setItem('configSimpleos',JSON.stringify(payload));
+(async () => {
 
-	platformBrowserDynamic().bootstrapModule(AppModule, {
-		preserveWhitespaces: false
-	}).catch(err => console.log(err));
-});
+    if (!localStorage.getItem('simplEOS.activeChainID')) {
+        localStorage.setItem('simplEOS.activeChainID', 'aca376f206b8fc25a6ed44dbdc66547c36c6c33e3a119ffbeaef943642f0e906');
+    }
 
+    const url = 'https://raw.githubusercontent.com/eosrio/simpleos/master/config.json';
+
+    let response;
+    try {
+        response = await fetch(url);
+    } catch (e) {
+        console.log('failed to load updated config.json from github');
+        console.log(e);
+    }
+
+    let jsonBody;
+    try {
+        jsonBody = await response.json();
+    } catch (e) {
+        console.log('error parsing json data');
+        console.log(e);
+    }
+
+    try {
+        if (jsonBody) {
+            const payload = {lastUpdate: new Date(), config: jsonBody};
+            localStorage.setItem('configSimpleos', JSON.stringify(payload));
+        }
+    } catch (e) {
+        console.log('error saving to localStorage');
+    }
+
+    // Launch Main Application
+    angularBoot();
+})();
